core/auth/requirements.txt
fastapi
uvicorn
redis[asyncio]
bcrypt
PyJWT
